Fix HF seed bug
Setting the HF seed at the end resulted in all jobs getting the same seed value. This seed value is taken from root_params.yaml, or set to 0 if not available. This means that all realisations got the same seed, or all had a random seed, even if one was set in sim_params.yaml. This fix means that if a seed is available it will be used as it will be caught by the for loop, while realisations without an explicit seed will still use the shared/0 value as it did previously.